// web/src/generated/models/matterSavedSearchRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Priv-View Core API
 * Phase-one Core API for authentication, hierarchical tenants, clients, matters, and matter metadata definitions.
 * OpenAPI spec version: 0.1.0
 */
import type { MatterSavedSearchReadVisibility } from './matterSavedSearchReadVisibility';
import type { MatterSavedSearchUserRead } from './matterSavedSearchUserRead';
import type { MatterSearchRequest } from './matterSearchRequest';

export interface MatterSavedSearchRead {
  id: string;
  matter_id: string;
  name: string;
  description: string | null;
  visibility: MatterSavedSearchReadVisibility;
  search: MatterSearchRequest;
  owner: MatterSavedSearchUserRead;
  shared_users: MatterSavedSearchUserRead[];
  is_owner: boolean;
  created_at: string;
  updated_at: string;
}